reminder/requirements.txt
line-bot-sdk==1.13.0
google-cloud-firestore==1.4.0
google-auth==1.6.2
